requirements.txt
mcp[cli]>=0.1.0
starlette>=0.27.0
uvicorn>=0.24.0
httpx>=0.25.0
python-dotenv>=1.0.0
agentpay-sdk>=0.1.0 